Trait rusttyc::Abstract

pub trait Abstract: Eq + Sized + Clone + Debug {
    type Err: Debug;
    type VariantTag: Debug + Clone + Copy + PartialEq + Eq;
    fn variant(&self) -> Option<Self::VariantTag>;

    fn unconstrained() -> Self;

    fn meet(&self, other: &Self) -> Result<Self, Self::Err>;

    fn variant_arity(tag: Self::VariantTag) -> usize;

    fn nth_child(&self, n: usize) -> &Self;

    fn from_tag(tag: Self::VariantTag, children: Vec<Self>) -> Self;

    fn is_unconstrained(&self) -> bool { ... }

    fn arity(&self) -> Option<usize> { ... }
}

The main trait representing types throughout the type checking procedure. It is bound to the type checker as the Value for the Key parameter. As such, it needs to implement EnaValue.

Associated Types

type Err: Debug

Represents an error during the meet of two abstract types.

type VariantTag: Debug + Clone + Copy + PartialEq + Eq

Loading content...
Required methods

fn variant(&self) -> Option<Self::VariantTag>

fn unconstrained() -> Self

Returns an unconstrained abstract type.

fn meet(&self, other: &Self) -> Result<Self, Self::Err>

Computes the meet of two abstract values.

fn variant_arity(tag: Self::VariantTag) -> usize

fn nth_child(&self, n: usize) -> &Self

fn from_tag(tag: Self::VariantTag, children: Vec<Self>) -> Self

Loading content...
Provided methods

fn is_unconstrained(&self) -> bool

Determines if an element is unconstrained.

fn arity(&self) -> Option<usize>

Loading content...